wrap api in exposed root module; move api exlectron.ipcRenderer functions
